[NO-JIRA] Refactor API utils
