+++ .gitignore
@@ -21,7 +21,7 @@ yarn-error.log
 build
 app/helpers
 helpers/helpers
-server/api/helpers
+server/helpers
 
 # runtime stuff
 server/server_content/greasemonkey
