+++ docs/API structure.md
@@ -9,4 +9,4 @@ The actions (Promises) which the server uses to perform each capability are adde
 API requests are rate-limited by delaying requests to the API if the limit of 100 requests/hour is reached.
 When an API request comes in, [`server/index.js`](https://github.com/georgegillams/georgegillams.co.uk/blob/master/server/index.js) will look for an action matching the path and the http method for the action. If a matching API capability is found, the associated action (Promise) is resolved.
 
-If an API action (Promise) resolves and object, it will be returned to the client. If an API action (Promise) throws a [`CategorisedError`](https://github.com/georgegillams/georgegillams.co.uk/blob/master/server/api/utils/errors.js) then the server will sent back the error along with a suitable http status code.
+If an API action (Promise) resolves and object, it will be returned to the client. If an API action (Promise) throws a [`CategorisedError`](https://github.com/georgegillams/georgegillams.co.uk/blob/master/server/utils/errors.js) then the server will sent back the error along with a suitable http status code.
