+++ package.json
@@ -17,14 +17,14 @@
     "build:aws": "AWS=true npm run build",
     "build:test": "PORT=9001 PROJECT_UNDER_TEST=true npm run build:client && PORT=9001 PROJECT_UNDER_TEST=true npm run build:server",
     "build:clean": "rimraf ./build",
-    "build:client": "cross-env NODE_PATH=./server/api:./app NODE_ENV=production BUILT_AT=$( date -u +%s ) webpack --config config/webpack.prod.babel.js --color -p --progress --hide-modules --display-optimization-bailout",
+    "build:client": "cross-env NODE_PATH=./server:./app NODE_ENV=production BUILT_AT=$( date -u +%s ) webpack --config config/webpack.prod.babel.js --color -p --progress --hide-modules --display-optimization-bailout",
     "build:server": "cross-env NODE_ENV=development babel -d ./build/server ./server -s --copy-files",
     "clean": "npm run test:clean && npm run build:clean",
     "dev": "redis-server & cross-env SECRET_API_KEY=asdfghjkl MONZO_ACCESS_PASSWORD=asdfghjkl npm run start:dev",
     "fix:adapter-core-5": "cp ./scripts/adapter_core5.js node_modules/webrtc-adapter/src/js/adapter_core5.js",
     "fix:package-lock": "sed -i '' -e 's/http:\\/\\//https:\\/\\//g' $(find . -name 'package-lock.json')",
-    "link-helpers:clean": "rm -f $(pwd)/app/helpers $(pwd)/server/api/helpers",
-    "link-helpers": "npm run link-helpers:clean && ln -Ffs $(pwd)/helpers $(pwd)/app/helpers && ln -Ffs $(pwd)/helpers $(pwd)/server/api/helpers",
+    "link-helpers:clean": "rm -f $(pwd)/app/helpers $(pwd)/server/helpers",
+    "link-helpers": "npm run link-helpers:clean && ln -Ffs $(pwd)/helpers $(pwd)/app/helpers && ln -Ffs $(pwd)/helpers $(pwd)/server/helpers",
     "lint": "npm run lint:js && npm run lint:scss",
     "lint:js": "eslint . --ext .js,.jsx",
     "lint:js:server-only": "eslint server --ext .js,.jsx",
@@ -36,10 +36,10 @@
     "prettier": "prettier --config .prettierrc --write \"**/*.js\"",
     "start": "npm run start:prod",
     "start:aws": "npm run start:prod-aws",
-    "start:dev": "cross-env NODE_PATH=./server/api:./app NODE_ENV=development BUILT_AT=$( date -u +%s ) nodemon --ignore 'server/server_content' --ignore 'app/*' --exec babel-node --inspect server",
-    "start:prod": "cross-env NODE_PATH=./build/server/api NODE_ENV=production node ./build/server",
-    "start:prod-aws": "cross-env AWS=true NODE_PATH=./build/server/api NODE_ENV=production node ./build/server",
-    "start:test": "cross-env PORT=9001 NODE_PATH=./build/server/api NODE_ENV=production PROJECT_UNDER_TEST=true REDIS_MOCK=true node ./build/server",
+    "start:dev": "cross-env NODE_PATH=./server:./app NODE_ENV=development BUILT_AT=$( date -u +%s ) nodemon --ignore 'server/server_content' --ignore 'app/*' --exec babel-node --inspect server",
+    "start:prod": "cross-env NODE_PATH=./build/server NODE_ENV=production node ./build/server",
+    "start:prod-aws": "cross-env AWS=true NODE_PATH=./build/server NODE_ENV=production node ./build/server",
+    "start:test": "cross-env PORT=9001 NODE_PATH=./build/server NODE_ENV=production PROJECT_UNDER_TEST=true REDIS_MOCK=true node ./build/server",
     "start:production": "npm run test && npm run build && npm run start:prod",
     "backstopjs:test": "npm run start:test & (sleep 5 && node ./scripts/backstopjs/backstop.js)",
     "backstopjs:test:allow-failure": "npm run start:test & (sleep 5 && node ./scripts/backstopjs/backstop.js --allowFailure)",
