+++ server/utils/hash.js
