+++ server/utils/mapPathToAction.js
