+++ server/utils/setContentLastUpdatedTimestamp.js
@@ -1,6 +1,6 @@
-import { datumLoad, datumCreate, datumUpdate } from '../actions/datum';
-
 import lockPromise from 'utils/lock';
+
+import { datumLoad, datumCreate, datumUpdate } from '../api/actions/datum';
 
 export default function setContentLastUpdatedTimestamp() {
   let newContentUpdateData = {};
