+++ server/utils/userOwnsResource.js
@@ -1,4 +1,4 @@
-import { datumLoad } from '../actions/datum';
+import { datumLoad } from '../api/actions/datum';
 
 import { find } from './find';
